+++ src/helpers/objects.ts
@@ -1,32 +1,30 @@
 /**
  * Get the nested properties of an object
- * This solution is lighter than the lodash get-version and works fine for the translations.
+ * This solution is lighter than the lodash get-version.
  * Source: http://stackoverflow.com/a/6491621/6942210
  */
 export const getObjectPropertyValue = (obj: Object, path: string) => {
-    // convert indexes to properties
-    path = path.replace(/\[(\w+)\]/g, '.$1');
-
-    // strip a leading dot
-    path = path.replace(/^\./, '');
-
-    // separate paths in array
-    let pathArray = path.split('.');
+    const pathArray = path
+        .replace(/\[(\w+)\]/g, '.$1') // convert indexes to properties
+        .replace(/^\./, '') // strip a leading dot
+        .split('.'); // separate paths in array
 
     /** Avoid errors in the getValue function. */
     const isObject = (object) => {
         return object === Object(object);
     };
 
+    let result = JSON.parse(JSON.stringify(obj));
+
     for (let i = 0; i < pathArray.length; ++i) {
-        let k = pathArray[i];
-        if (isObject(obj) && k in obj) {
-            obj = obj[k];
+        const k = pathArray[i];
+        if (isObject(result) && k in result) {
+            result = result[k];
         } else {
             return;
         }
     }
-    return obj;
+    return result;
 };
 
 /**
@@ -36,19 +34,14 @@ export const getObjectPropertyValue = (obj: Object, path: string) => {
  * @param value Value to be set for the given property
  * Source: https://stackoverflow.com/a/13719799/6942210
  */
-export const setObjectPropertyValue = (obj: Object, path, value) => {
+export const setObjectPropertyValue = (obj: Object, path: string | string[], value: any) => {
     if (typeof path === 'string') {
         path = path.split('.');
     }
 
     if (path.length > 1) {
-        let e = path.shift();
-        setObjectPropertyValue(obj[e] =
-            Object.prototype.toString.call(obj[e]) === '[object Object]'
-                ? obj[e]
-                : {},
-            path,
-            value);
+        const e = path.shift();
+        setObjectPropertyValue(obj[e] = Object.prototype.toString.call(obj[e]) === '[object Object]' ? obj[e] : {}, path, value);
     } else {
         obj[path[0]] = value;
     }