+++ src/icons/folderIcons.ts
@@ -181,6 +181,8 @@ export const folderIcons: FolderTheme[] = [
             { name: 'folder-nuxt', folderNames: ['nuxt', '.nuxt'], enabledFor: [IconPack.Vuex, IconPack.Vue] },
             { name: 'folder-vue-directives', folderNames: ['directives'], enabledFor: [IconPack.Vuex, IconPack.Vue] },
             { name: 'folder-vue', folderNames: ['components'], enabledFor: [IconPack.Vuex, IconPack.Vue] },
+            { name: 'folder-mobile', folderNames: ['mobile', 'mobiles', 'portable'] },
+            { name: 'folder-stencil', folderNames: ['.stencil'] },
             { name: 'folder-rivescript', folderNames: ['rivescript', 'brain'] },
         ]
     },