+++ src/scripts/icons/checks/checkIconAvailability.ts
@@ -111,7 +111,7 @@ const handleErrors = () => {
     if (amountOfErrors > 0) {
         console.log('> Material Icon Theme:', painter.red(`Found ${amountOfErrors} error(s) in the icon configuration!`));
     } else {
-        console.log('> Material Icon Theme:', painter.green(`Passed icon availability checks!`));
+        console.log('> Material Icon Theme:', painter.green('Passed icon availability checks!'));
     }
     logIconInformation(wrongIconNames.fileIcons, 'File icons');
     logIconInformation(wrongIconNames.folderIcons, 'Folder icons');